requirements.txt
langchain
langgraph
langchain-community
langchain-openai
langchain-pinecone
openai
pinecone
pypdf
unicorn
fastapi
python-dotenv

